# apps/dashboard/lib/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Carbon API",
    "description": "Control plane for Carbon — projects, ingestion, emulators, snapshots, artifacts, org/billing, and CLI auth. Every endpoint accepts either an API key (`x-carbon-key`) or a Better Auth session cookie/bearer token; the docs and `/health` endpoints are the only routes served anonymously.",
    "version": "dev",
    "contact": {
      "name": "Carbon",
      "url": "https://github.com/TaxCollector23/carbon",
      "email": "support@example.com"
    },
    "license": { "name": "Apache-2.0", "url": "https://www.apache.org/licenses/LICENSE-2.0" }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-carbon-key",
        "description": "A key minted via `POST /v1/api-keys` or the `bootstrap` script. Prefix `ck_live_*` for production, `ck_test_*` for test envs."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "better-auth-session",
        "description": "A Better Auth session bearer token. Preferred for programmatic access from a signed-in user."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "The Better Auth session cookie set by the dashboard. Used implicitly when the browser calls the API cross-origin with `credentials: \"include\"`."
      }
    },
...
  },
  "paths": {
...
    "/v1/emulators/{id}/snapshot": {
      "post": {
        "summary": "Snapshot an emulator",
        "tags": ["Emulators"],
        "description": "Persist the emulator's current state under `name`. Overwrites any existing snapshot with the same name.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[a-z0-9][a-z0-9-_]*$"
                  }
                },
                "required": ["name"],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "name": { "type": "string" } },
                  "required": ["name"],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{slug}/snapshots": {
      "get": {
        "summary": "List project snapshots",
        "tags": ["Snapshots"],
        "description": "Enumerate saved snapshots for a project. Bounded scan — pass `limit` (max 500) to stop early; `truncated: true` means more snapshots exist beyond the limit.",
        "parameters": [
          {
            "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 },
            "in": "query",
            "name": "limit",
            "required": false
          },
          { "schema": { "type": "string" }, "in": "path", "name": "slug", "required": true }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "size": { "type": "integer" },
                          "modifiedAt": { "type": "number" }
                        },
                        "required": ["name", "size", "modifiedAt"],
                        "additionalProperties": false
                      }
                    },
                    "limit": { "type": "integer" },
                    "truncated": { "type": "boolean" }
                  },
                  "required": ["data", "limit", "truncated"],
                  "additionalProperties": false
                },
                "example": {
                  "data": [
                    { "name": "seed-cart-full", "size": 4212, "modifiedAt": 1731601361000 },
                    { "name": "post-checkout", "size": 5893, "modifiedAt": 1731597811000 }
                  ],
                  "limit": 100,
                  "truncated": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/snapshots": {
      "post": {
        "summary": "Save a snapshot",
        "tags": ["Snapshots"],
        "description": "Persist a `StateSnapshot` for the given project under `name`. Overwrites any existing snapshot with the same name; name must match /^[a-z0-9][a-z0-9-]{0,63}$/ since it is used as a storage key segment.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectSlug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[a-z0-9][a-z0-9-]*$"
                  },
                  "name": { "type": "string", "pattern": "^[a-z0-9][a-z0-9-]{0,63}$" },
                  "snapshot": {
                    "type": "object",
                    "properties": {
                      "version": { "type": "number", "enum": [1] },
                      "takenAt": { "type": "number" },
                      "records": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "resource": { "type": "string" },
                            "id": { "type": "string" },
                            "data": { "type": "object", "additionalProperties": {} },
                            "createdAt": { "type": "number" },
                            "updatedAt": { "type": "number" }
                          },
                          "required": ["resource", "id", "data", "createdAt", "updatedAt"],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": ["version", "takenAt", "records"],
                    "additionalProperties": false
                  }
                },
                "required": ["projectSlug", "name", "snapshot"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "storageKey": { "type": "string" }
                  },
                  "required": ["name", "storageKey"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{slug}/snapshots/{name}": {
      "get": {
        "summary": "Get /v1/projects/:slug/snapshots/:name",
        "tags": ["Snapshots"],
        "description": "GET /v1/projects/:slug/snapshots/:name",
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "slug", "required": true },
          { "schema": { "type": "string" }, "in": "path", "name": "name", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Delete /v1/projects/:slug/snapshots/:name",
        "tags": ["Snapshots"],
        "description": "DELETE /v1/projects/:slug/snapshots/:name",
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "slug", "required": true },
          { "schema": { "type": "string" }, "in": "path", "name": "name", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/snapshots/{slug}/diff": {
      "get": {
        "summary": "Get /v1/snapshots/:slug/diff",
        "tags": ["Snapshots"],
        "description": "GET /v1/snapshots/:slug/diff",
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "slug", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
  },
  "servers": [{ "url": "http://localhost:4000", "description": "local dev" }],
  "security": [{ "apiKey": [] }, { "bearerAuth": [] }, { "sessionCookie": [] }],
...
}